+++ src/components/dialog/dialog.style.js
@@ -64,7 +64,6 @@ const DialogStyle = styled.div`
   }
 
   ${StyledFormContent}.sticky {
-    height: calc(100vh - 210px);
     padding-right: ${HORIZONTAL_PADDING}px;
     padding-left: ${HORIZONTAL_PADDING}px;
     padding-top: 20px;