+++ src/components/message/message-content/message-content.style.js
@@ -4,6 +4,7 @@ import { margin } from "styled-system";
 const MessageContentStyle = styled.div`
   padding: 15px 50px 15px 20px;
   white-space: pre-wrap;
+  flex: 1;
 
   .carbon-content__title {
     margin-bottom: 2px;