+++ src/components/pages/page/page.spec.js
@@ -5,22 +5,23 @@ import Page from "./page.component";
 import FullScreenHeading from "../../../__internal__/full-screen-heading";
 
 describe("Pages", () => {
+  const wrapper = shallow(
+    <Page
+      transitionName={() => {}}
+      title="My Title"
+      data-element="carbon-page-content"
+    >
+      My Content
+    </Page>
+  );
+
   it("renders a page with a full screen heading", () => {
-    const wrapper = shallow(
-      <Page
-        transitionName={() => {}}
-        title="My Title"
-        data-element="carbon-page-content"
-      >
-        My Content
-      </Page>
-    );
-    const fullScrenHeading = wrapper.find(FullScreenHeading);
+    const fullScreenHeading = wrapper.find(FullScreenHeading);
     expect(wrapper.find(StyledPage).props()["data-element"]).toEqual(
       "carbon-page-content"
     );
     expect(wrapper.find(StyledPage).props()["data-component"]).toEqual("page");
-    expect(fullScrenHeading.props().children).toEqual("My Title");
+    expect(fullScreenHeading.props().children).toEqual("My Title");
     expect(
       wrapper.find(StyledPage).props().children[1].props.children.props.children
     ).toEqual("My Content");